Move FetchPlugin in a dedicated directory
