+++ qubesbuilder/plugins/fetch/scripts/download-and-verify-file
